chore: added missing setup
