+++ crates/amaru/src/sync/pull.rs
@@ -12,13 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use amaru_kernel::Point;
 use amaru_ouroboros::protocol::{peer::PeerSession, PullEvent, RawHeader};
 use gasket::framework::*;
 use miette::miette;
-use pallas_network::miniprotocols::{
-    chainsync::{HeaderContent, NextResponse, Tip},
-    Point,
-};
+use pallas_network::miniprotocols::chainsync::{HeaderContent, NextResponse, Tip};
 use pallas_traverse::MultiEraHeader;
 use std::time::Duration;
 use tokio::time::timeout;
@@ -79,9 +77,16 @@ impl Stage {
     pub async fn find_intersection(&self) -> Result<(), WorkerError> {
         let mut peer_client = self.peer_session.peer_client.lock().await;
         let client = (*peer_client).chainsync();
-
+        fn new_point(point: &Point) -> pallas_network::miniprotocols::Point {
+            match point.clone() {
+                Point::Origin => pallas_network::miniprotocols::Point::Origin,
+                Point::Specific(slot, hash) => {
+                    pallas_network::miniprotocols::Point::Specific(slot, hash)
+                }
+            }
+        }
         let (point, _) = client
-            .find_intersect(self.intersection.clone())
+            .find_intersect(self.intersection.iter().map(new_point).collect())
             .await
             .or_restart()?;
 
@@ -178,7 +183,15 @@ impl gasket::framework::Worker<Stage> for Worker {
                 stage.roll_forward(&header).await?;
             }
             NextResponse::RollBackward(point, tip) => {
-                stage.roll_back(point, tip).await?;
+                fn new_point(point: pallas_network::miniprotocols::Point) -> Point {
+                    match point.clone() {
+                        pallas_network::miniprotocols::Point::Origin => Point::Origin,
+                        pallas_network::miniprotocols::Point::Specific(slot, hash) => {
+                            Point::Specific(slot, hash)
+                        }
+                    }
+                }
+                stage.roll_back(new_point(point), tip).await?;
             }
             NextResponse::Await => {}
         };
