+++ crates/kernel/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "amaru-ledger"
+name = "amaru-kernel"
 version.workspace = true
 edition.workspace = true
 description.workspace = true
@@ -11,19 +11,15 @@ documentation.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+bech32.workspace = true
 hex.workspace = true
 num.workspace = true
+pallas-addresses.workspace = true
+pallas-codec.workspace = true
+pallas-crypto.workspace = true
+pallas-primitives.workspace = true
 serde.workspace = true
-tracing.workspace = true
 thiserror.workspace = true
-
-amaru-kernel = { path = "../kernel" }
 
 [dev-dependencies]
 proptest.workspace = true
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[target.'cfg(not(std))'.dependencies]
-num = { version = "0.4.3", default-features = false, features = ["alloc", "libm"] }
