+++ crates/kernel/src/iter/borrow.rs
