+++ crates/ledger/src/state/volatile_db.rs
@@ -13,12 +13,10 @@
 // limitations under the License.
 
 use super::{diff_bind::DiffBind, diff_epoch_reg::DiffEpochReg, diff_set::DiffSet};
-use crate::{
-    kernel::{
-        epoch_from_slot, Epoch, Lovelace, Point, PoolId, PoolParams, StakeCredential,
-        TransactionInput, TransactionOutput,
-    },
-    store::{self, columns::*},
+use crate::store::{self, columns::*};
+use amaru_kernel::{
+    epoch_from_slot, Epoch, Lovelace, Point, PoolId, PoolParams, StakeCredential, TransactionInput,
+    TransactionOutput,
 };
 use std::collections::{BTreeSet, VecDeque};
